Transform: Rename transform_dead_code -> transform_remove_code
