+++ loki/transform/__init__.py
@@ -19,6 +19,6 @@ from loki.transform.build_system_transform import * # noqa
 from loki.transform.transform_hoist_variables import * # noqa
 from loki.transform.transform_parametrise import * # noqa
 from loki.transform.transform_extract_contained_procedures import * # noqa
-from loki.transform.transform_dead_code import * # noqa
+from loki.transform.transform_remove_code import * # noqa
 from loki.transform.transform_sanitise import * # noqa
 from loki.transform.pipeline import * # noqa
